# Speed up compiles and clean up RUSTFLAGS