--- a/.cargo/config.toml
+++ b/.cargo/config.toml
@@ -8,7 +8,33 @@ rustflags = [
     "-C",
     "panic=abort",
     "-C",
-    "target-feature=+relax,+unaligned-scalar-mem,+b",
+    "target-feature=+relax,+unaligned-scalar-mem,+zba,+zbb,+zbc,+zbs",
     "-C",
     "force-frame-pointers=no",
+    # See https://github.com/tock/tock/pull/2853
+    "-C",
+    "relocation-model=static",
+    # Opt-in to Rust v0 symbol mangling scheme.
+    #   See https://github.com/rust-lang/rust/issues/60705 and
+    #   https://github.com/tock/tock/issues/3529.
+    "-C",
+    "symbol-mangling-version=v0",
+    "-C",
+    # Tell rustc to use the LLVM linker. This avoids needing GCC as a
+    # dependency to build the kernel.
+    "linker=rust-lld",
+    "-C",
+    # Use the LLVM lld executable with the `-flavor gnu` flag.
+    "linker-flavor=ld.lld",
+    "-C",
+    # lld by default uses a default page size to align program
+    # sections. Tock expects that program sections are set back-to-back. `-nmagic`
+    # instructs the linker to not page-align sections.
+    "link-arg=-nmagic",
+    # Identical Code Folding (ICF) set to all. This tells the linker to be
+    # more aggressive about removing duplicate code. The default is `safe`, and
+    # the downside to `all` is that different functions in the code can end up with
+    # the same address in the binary. However, it can save a fair bit of code  size.
+    "-C",
+    "link-arg=-icf=all",
 ]