--- a/runtime/src/board.rs
+++ b/runtime/src/board.rs
@@ -4,8 +4,6 @@ use crate::chip::VeeRDefaultPeripherals;
 use crate::chip::TIMERS;
 use crate::timers::InternalTimers;
 use capsules_core::virtualizers::virtual_alarm::{MuxAlarm, VirtualMuxAlarm};
-use capsules_runtime::mctp::mux::MuxMCTPDriver;
-use capsules_runtime::mctp::transport_binding::MCTPI3CBinding;
 use core::ptr::{addr_of, addr_of_mut};
 use kernel::capabilities;
 use kernel::component::Component;